refactor: move task declaration to a new file
