+++ src/executor.c
@@ -1,4 +1,5 @@
 #include "executor.h"
+#include "task.h"
 
 #include "stdatomic.h"
 #include "stdio.h"
