+++ src/node.c
@@ -1,7 +1,6 @@
 #include "node.h"
 
-#include "data_container.h"
-#include "executor.h"
+#include "task.h"
 #include "graph.h"
 #include "instance.h"
 #include "stdatomic.h"
